refactored structure src/ => src/lib/ and its footprint
