+++ src/lib/sign-test.js
