+++ src/lib/string-polyfills.js
