+++ src/lib/trim.js
